uClibc patch: (refs #842) Enable IPv6 for Linux kernel smaller 2.6.14
This is a refreshed version of 011-ipv6_oldkernels.patch by Whoopie.

The patch was renamed to clearly indicate its function (as suggested by er13).

An identical include was removed and the include moved to top of file.
Some comments were added and the code was beautified a bit.

Tested IPv6 with my Speedport W 701V box based on v2.6.13.1 Linux-kernel.

Signed-off-by: Sedat Dilek <sedat.dilek@gmail.com>

git-svn-id: file:///var/svn/freetz/trunk@7269 149334a1-2f27-0410-a3b9-fc62619ac1e6