''fwmod'' speed optimisation:   - If the current firmware image (incl. alien) has been unpacked before, do     not unpack it again to save some time.   - If the firmware + alien config combination has changed, delete the old     ''.unpacked_*'' marker file.   - Marker files now contain the names of the two firmware images as suffixes     to uniquely identify the previous unpack action.
'''Hint:''' If the unpack code in this script changes in the future and users
have problems because the old unpack action for the same firmware does not do
exactly what is expected, just tell them to call `make common-clean` or
`rm -f build/.unpacked_*` to force a new unpack.
You may also just switch the box or alien type for the same effect.



git-svn-id: file:///var/svn/freetz/trunk@8443 149334a1-2f27-0410-a3b9-fc62619ac1e6