In r8569 we saw a problem with the new version of `echoX` due to Bash ''getopts'' argument handling. Thanks ralf for fixing the problem in ''tools/freetz_patch'' earlier (refs #1700).
Anyway I made the function `echoX` a little smarter so it would be able to cope
with most cases of parameters like "----", "-blah" etc. Reasons and exceptions
are documented in a rather lengthy source code comment now. Please not that in
the future there will be no more ''getopts'' errors printed to the console
because I switched to silent error handling mode. But if strange stuff gets
printed to the console in the future, we know how to handle it. I just checked
the code base for other problematic parts, but there are none others.


git-svn-id: file:///var/svn/freetz/trunk@8571 149334a1-2f27-0410-a3b9-fc62619ac1e6